services/agents/requirements.txt
anthropic==0.7.8
yfinance==0.2.36
pandas==2.2.1
numpy==1.26.4
aiohttp==3.9.3
python-dotenv==1.0.1
multi-agent-orchestrator==0.1.0
fastapi>=0.100.0
uvicorn>=0.23.1
pydantic>=2.0.3
python-dotenv==1.0.1